Lots of changes to support usage
Due to the size of the change, I don't expect these changes to be merged
upstream. But a number of them seem essential for actually using the package, so
who knows.

(1) Export some types

In order to support dashboard declaration using the struct literal syntax,
and just to set some properties at all, I exported a number of types:

Board:
- Templating
- Annotation
- Time
- Timepicker
- TemplateVar
- Option
- Current

Panel:
- SeriesOverride
- Tooltip
- Axis

(2) Add omitempty tag to a number of non-required settings (prometheus)

(3) Add the Target member to support Graphite

(4) Use io.Reader as the parameter type to doRequest.

Without that, it panics for GETs otherwise due to a nil-but-not-nil interface.

(5) Concatenate the query string on the path rather than replacing it.

This supports base URLs other than a bare hostname

(6) Remove board member from Row type.

The circular reference made it impossible to declare Boards and Rows using a
struct literal.