Increase the limit of users returned
As per documentation, `/api/users` returns only the first 1000 users. This means, that for Grafana instances that have more than 1000 users, this SDK function does not return all users.

One option could be introducing a parameter, for `GetAllUsers()`, but that would be a breaking change for all existing users. As an alternative, I suggest setting the limit to a very high number (`99999`, in this case). This would avoid a breaking change and would make `GetAllUsers()` to work as expected (since the name implies it returns ALL users).