Add alertnotification API support (#53)
* datasource: add secureJsonData

Add a new member to the `Datasource` struct called SecureJSONData which
gets passed as `secureJsonData`. This is needed because certain features
are only available via that e.g. passing custom header values.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* Add alertnotification API wrappers

Add `alertnotification.go` which contains the struct definition of an
`alertnotification`. Add `rest-alertnotification.go` which contains the
API wrappers. Edit `panel.go` to remove the old definition of an
alertnotification since it is not complete and it belongs in its own
file.

Signed-off-by: Giedrius Statkevičius <giedriuswork@gmail.com>

* rest-alertnotification: fix PUT call

It is actually simply `/api/alert-notifications/:id`