Fix GetDashboard API mess, update README (#73)
* rest: dashboard: separate slug/uid calls

Separate UID/slug calls of `/api/dashboards/...` into two different
methods. Hide the general methods which are called by the wrappers. This
will allow us in the future to remove the slug calls once they get
removed from Grafana.

Remove the second variable returned by `setPrefix` because it is not
used. Remove the `uid/` handling from there as it should not know about
such cases. It is merely for adding a prefix `file/` or `db/` depending
on the passed string. `uid` should be moved into the separate method.

Fix the `Deprecated` comment format so that it would show up in IDEs
properly.

This overall improves the clarity of the API that we expose to our
users. Finally, lets reuse getRawDashboard() in the other method which
returns the same data, just parsed.

* Update README

* cmd: backup-dashboards: fix to use the new API

* rest: dashboard: remove extra setPrefix() call

* test: mark helpers, separate tests into proper files

* test: fix typos, Error() -> Fatal()

It is really impossible to continue if any of those calls fail.