sdk: add context.Context to REST calls
Add `context.Context` as the first argument to functions which do
requests. This allows users to have fine-grained control over the
timeouts in the typical Go manner - they will not have to have the same
timeout on all operations that is set via `http.Client`.

This also informs our users that these functions do requests and that
they *do* block for a good reason - a HTTP request is made and then we
block until a response comes.