user: Allow signed in user to change organization
Adds a new method to allow the currently signed user to switch its
context into a given organization (provided via the organizationId).
This is particularly useful when trying to automate processes like
generating tokens for a given organization as explained in:
https://grafana.com/docs/grafana/latest/tutorials/api_org_token_howto/.

This PR reflects the /api/user/using/:organizationId API call.