Support object-typed datasource references
See: https://github.com/grafana/grafana/pull/33817

Since Grafana v8.4.3 datasource references are not longer a string with
the datasource name, but an object with the type and UID instead.

This means that dashboards saved with Grafana v8.4.3 or higher are not
longer parseable by this SDK.

I replaced the `*string` datasource by a `DatasourceRef` struct, which
has custom unmarshaling: it tries to unmarshal the new object, and if it
fails, it will try to unmarshal data as string and put it into
`LegacyName` instead.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>